backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
numpy==1.26.4
pandas==2.2.2
requests==2.32.3
joblib==1.4.2
scikit-learn==1.5.1
google-api-python-client==2.143.0
google-auth==2.35.0
python-multipart==0.0.9
